Add types.d.ts and change type imports on files in the SearchResults folder
