+++ src/utils/helpers/formatPascalCase.ts
@@ -1,4 +1,4 @@
-export default function formatSellerName(name: string): string {
+export default function formatPascalCase(name: string): string {
   return name
     .split(' ')
     .map(e => e[0].toUpperCase() + e.slice(1).toLowerCase())
