+++ src/views/SearchResults/components/DropdownFilterOptions/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DropdownFilterOptions'
