+++ src/views/SearchResults/components/DropdownFilters/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DropdownFilters';
